[table-driven-branch] Create some basic tests for table-driven protos.
During development, we need a way to verify that the generator and runtime
changes continue to compile and behave sensibly, even if we don't have
every single feature built out yet. To this end, build a test suite
around the large unittest.proto that we can run independently as we
build out new functionality.