[table-driven-branch] Initial definition of table-driven message layout and in-memory storage. (#1861)
This is the first chunk of work driving toward table-driven messages.
The best way to view this would be commit-by-commit, rather than the
whole PR diff. I've opened a separate `table-driven-branch` for all this
work so that I can iterate without interfering with `main` development.

As discussed offline, the general idea is that we do something similar
to what we did for `NameMap` by introducing a `StaticString`-encoded
description that compactly represents the layout of the message. By
combining this with a single common storage class that manages a raw
memory buffer based on that layout, we can dramatically reduce codegen
size (we're no longer generating unique metadata for each private
storage class, and eventually we won't be generating unique methods for
decode, traverse, `==`, and so forth).

This PR introduces the following concepts:

* `_MessageLayout`, a helper that extracts data from the `StaticString`
description for a particular message type
* `_MessageStorage`, the common storage class that allocates space and
manages slices of it based on a `_MessageLayout`. It contains
accessors/mutators for fields based on their byte offset, and eventually
will house the other fundamental operations like traverse, decode,
equality, and so forth.
* The generator is updated to produce the new implementation for
accessors and the message layout string, and I've carved out a tiny
subset of the unittest.proto files to generate them in isolation so that
I can verify that the generated code compiles and I can write some basic
unit tests as I build out more functionality.

There are still a lot of missing parts here, like maps, oneofs,
extensions, etc.