[table-driven-branch] Implement deinitialization and copying. (#1868)
Note: `oneof`s and `map`s are still not implemented.

For most cases, deinitialization of message storage is fairly
straightforward. For scalar trivial values, we needn't do anything. For
nontrivial types like `String`, `Data`, and arrays of fields, we have to
tell the compiler statically what the actual type is so that it can
generate the right code for the operation.

Copying message storage (to implement CoW) is similar. We ignore scalar
values at first, because they (along with has-bits) are stored together
at the front of the buffer. While we iterate over the fields, we
copy-initialize any non-trivial fields (again telling the compiler what
the actual type is), and we record the offset of the first non-trivial
field we saw. That lets us do a single "memcpy" at the end to copy all
the trivial fields. This is a small optimization but it seems worthwhile
to take advantage of this characteristic of the layout.

> NOTE: There are still be some cases that we could optimize. For
example, a message with only trivial fields would still iterate over all
of the field layouts, do nothing in the loop, and then do nothing else
(for deinit) or just do a memcpy (for copying). So it might be
worthwhile to use a bit somewhere in the layout to describe whether a
message is only trivial fields. Leaving this for a future PR.

What is less straightforward: we need a way to deal with arbitrary
message types (singular and repeated). The original plan was to use the
generated submessage accessor function for this, but that doesn't work
well for arrays of messages. After chatting with @thomasvl offline, I'm
going to look at replacing the submessage accessor concept with a set of
"submessage operations" that would be a sort of callback where
`_MessageStorage` can delegate an operation on a value of an arbitrary
complex type to the message, who would have more information generated
about the types in question.