[table-driven-branch] Deinitialize and copy submessage fields. (#1869)
In order to perform the low-level typed operations on the raw memory,
the protobuf runtime needs to be able to refer statically to the type of
a submessage field. This is inherently challenging when we're trying to
use a single storage class to represent all possible messages.

To achieve this, we generate trampoline functions that are passed into
the `_MessageLayout` initializer and stored. Then, for example, when
deinitializing a submessage field, the storage class calls the
deinitializer trampoline with the appropriate submessage index and the
trampoline function immediately calls back into the storage class but
passes it the type hint so that the memory can be bound to the correct
type. Likewise for copying a submessage field.

This replaces the original notion of a submessage accessor function that
relied on returning the metatype as an existential value. That wouldn't
be compatible with embedded Swift; these type hints should be.

I've also updated the generator to avoid generating these trampoline
functions if a message doesn't have any submessages. In this case, we
can just pass a placeholder function baked into the runtime that does a
precondition failure, so we save a bit on codegen.