[table-driven-branch] Add support for `oneof`s. (#1870)
This commit adds support for in-memory storage (and getters/setters) of
`oneof`s and their members.

We keep the same API as before: we generate an `enum` with associated
values to represent the `oneof` itself, and we also provide getters and
setters for the individual members directly on the containing message.

Though this is invisible to the end user, the implementation has
"flipped". In the current main branch, the value we store is the enum
value, and the member getters/setters extract the value from that. With
the table-driven approach, we store the values in memory as a `UInt32`
that tracks the currently set field along with whatever storage a field
of each member type would normally require. So the member accessors are
now "fast" and directly manipulate storage, while the enum accessor does
a `switch` to determine which field needs to be manipulated.

Possible future optimization: Today, we allocate separate storage in
memory for each oneof field, even though only one of
them can be set at any time. We could do what µpb does and have the
storage overlap (choosing the max size among all the member types).

Since this is the first time the binary-search subscript is being used
in `_MessageLayout`, I've also fixed a bug there where I had a
comparison reversed. I blame this on µpb naming both of their operands
`num` and `number`...