[table-driven-branch] Fix some copy/deinit bugs. (#1871)
For copying, we were computing the size of the bitwise-copied range
incorrectly. We were tracking the offset of the first non-trivial field
we found, but this isn't correct because we iterate over them in *field
number* order, not based on their *storage kind*. Instead, we need to
track the *lowest* offset. It may be worth it to store this in the
layout after all so that we don't have to recompute it each time.

For both copy and deinit, I had to fill in the missing oneof support.
It's surprising that the tests passed without this, but I think that was
by coincidence with the bug above, since it was causing us to
bitwise-copy *more* data than we otherwise would have, and things just
happened to line up correctly.