Update OneofGenerator to support breaking up large switch statements (#1866)
Given a oneof field with ~600 options, the generated traverse function
will fail to compile because swift is unable to check the switch is
exhaustive in reasonable time. See
https://github.com/apple/swift-protobuf/issues/1856

This PR follows similar logic to #904 which allows switch statements to
be broken up into multiple switch statements (using a `default: break`
case).

Instead of hardcoding 500, I've added a new GeneratorOption to support
setting the "maxCasesInSwitch" which allows the same number to be used
in `EnumGenerator` and `OneofGenerator` and allows clients to set this
number to whatever suits them based on Swift version or other
considerations. This number could also be set to something very high to
effectively turn off the feature.

The second commit adds tests, I asked claude to help out with this part
so please help me in double checking those tests. I was able to run
tests successfully locally.