Support `isInitialized`. (#1874)
This PR implements `isInitializedShallow` on the storage class, which is
the fast-path that just checks the has-bits using a bitwise-compare. The
public `isInitialized` is built on top of that—it does the shallow check
first, and if that succeeds, then it walks the layout and checks the
submessage fields recursively.

I'm growing a bit weary of how many of these "trampoline" functions I've
had to introduce. I think the `isInitialized` one could be redesigned to
be something that just returns a layout and a storage, but some of the
others need to be able to propagate explicit type hints through the API
layers. I'd like to find a way to generalize them that doesn't rely on
existentials/too many generics so that we don't risk designing something
that either doesn't work on embedded or gets code bloat from too many
explicit specializations, but I haven't come up with anything yet. Maybe
once I've implemented the full spread of what we need, an idea will come
to mind.