[table-driven-branch] Implement binary serialization.
This PR implements `serializedBytesSize()` and `serializedBytes()` on
`_MessageStorage`. Eventually, we'll be able to put the call to the
latter completely in the runtime (in `_MessageImplementationBase`),
but for now I'm generating the method directly because we still need
the old WKTs and plugin protos to work until we migrate those to be
table-driven.

I've also fixed a bug in the earlier `isInitialized` logic; we were
checking *any* message/group field to see if it was initialized,
including optional message/group fields. Fixed it to allow unset
optional messages/groups and only recurse into them when they're
present.

I've brought over the subset of the `TestAllTypes` encoding tests
that are able to pass today. This should cover scalar and repeated
fields (packed and unpacked), submessages/groups, and oneofs. I
also brought over the subset of required field encoding tests to
verify the expected throwing and partial behavior.