[table-driven-branch] Create a general trampoline function.
Some of the operations we're going to implement, like
`serializedBytesSize` and `serializedBytes`, really just need
a way to invoke an arbitrary operation on the underlying storage
of a submessage.

`isInitialized` can also be implemented on top of this. It means
that we don't short-circuit as quickly as we could, but I don't
think the minor performance boost of skipping a dispatch is worth
the complexity of having another trampoline function.