Build `protoc` from source (#1876)
## Motivation

In a recent addition, we included changes to provide `protoc` as a
binary target from the official releases from the upstream `protobuf`
repository. While this solved one of the largest issues in our ecosystem
where folks currently needed to install `protoc` through a system
package manager, it also introduced another set of problems that binary
targets bring with them.

## Modifications

This PR adds `protobuf` and `abseil-cpp` as a submodule to this project
and includes a new target that builds both together from source. It
includes the minimal set of files from both to successfully build
`protoc` and also passes the right compiler and linker flags to build
those warning free.

## Result

We are no longer using a binary target for `protoc` but instead build it
from source which alleviates the problems that we introduced with the
binary target.