[table-driven-branch] Implement binary decoding for singular message/group fields. (#1895)
This introduces a couple closure-based helper methods on
`WireFormatReader` that create sub-readers to parse submessages and
groups. The sub-readers track their own termination state and then the
parent reader is updated when the sub-reader is done.

In doing this, I removed the separate tracking of the maximum depth
limit and remaining recursion budget. By only exposing sub-reader based
APIs, we can just have a single budget that decreases each time a
sub-reader is created.

I've introduced the notion of an abstract "submessage operation" that
gets passed to `performOnSubmessageStorage` since we now need to handle
mutation (and not only mutation, but first-time creation if a field
isn't present). The plan is for this to also grow to support appending,
for decoding repeated fields.

Also fixed another small binary search bug for field layouts. µpb's
"denseBelow" actually means "denseAtMost", but I defined ours as
strictly below, so I was hitting a boundary bug that just happened to
coincide with the treatment of groups in our test protos.