Optimize the varint size calculations to be branchless. (#1905)
The logic for this comes from the C++ code:

https://github.com/protocolbuffers/protobuf/blob/195da42b90bd39aa8c0b6551188fccf77a83f506/src/google/protobuf/io/coded_stream.h#L1754-L1826

I've always thought Swift didn't have a way to get to the optimal
instructions, but I was wrong, it has all the building blocks needed.

Looking at new code in godbolt, in `-O`, without or without the
arithmetic overflow protection, the code is the same (tiny), but by
turning off the arithmetic overflow checks, we also keep the debug code
a bit smaller.