[table-driven-branch] Add decoding/encoding for enum fields.
As singular fields, we store the underlying Int32 value of the
enum in storage instead of the Swift enum value because it's
simpler. For repeated enums, since we don't want to penalize
the performance of accessors that return Swift arrays, we store
the `Array` of Swift enum values.

This means that we need trampoline functions for some specific
cases:

* Deinitializing and copying a repeated enum field.
* Checking whether two Swift enum arrays are equal.
* When encoding a repeated enum, we need to get the raw values.
* When decoding a singular or repeated enum, we need to validate
  the Int32s on the wire to make sure they're valid cases.

So this change is larger than I had hoped, because we needed to
introduce a new trampoline (`performOnRawEnumValues`), and then
it made sense to rename some of the existing trampoline
infrastructure that no longer just serves submessages.