[table-driven-branch] Update generator to produce layout strings for map entries.
To encode/decode map entries, we want to reuse as much of the existing
logic as possible. Since map entries are really just submessages with
two fixed fields, we can generate layout strings for them and use the
existing `_MessageStorage` serialization logic to manipulate them. We
don't want to generate full messages for them because this would be
mostly bloat; the layout strings, plus additional functionality to
come in the trampolines, should be enough to support what we need.