[table-driven-branch] Complete encoding/decoding implementation for maps.
The `performOnMapEntry` trampoline function does the translation between
message storage (laid out like a map entry on the wire) and the Swift
dictionary that's actually stored in memory. This required the creation
of some "proxy/participant" types to abstract out common behavior without
forcing multiple overloads of the function with different generic
constraints. This is similar to what we have today in FieldTypes.swift,
but smaller since those will be going away after the new implementation
is done.