Move the API check into a separate github CI config.
It was already running in parallel, this just makes it easier to make it not
required to deal with the fact that some upstream protos can pick up new enum
elements which isn't really breaking to SwiftProtobuf.

Remove the known breaks file that was used to land the last update.